requirements.txt
Flask==2.0.2
mysql-connector-python==8.0.29
python-dotenv==0.19.2
Werkzeug==2.0.3   # Downgrade to a compatible version

